remove rsonpath-test crate
