+++ crates/rsonpath-lib/.gitignore
@@ -1,5 +1,3 @@
-/target
-/Cargo.lock
 /tests/documents/json/*
 !/tests/documents/json/large
 /tests/documents/json/large/compressed
