+++ crates/rsonpath-lib/Cargo.toml
@@ -42,6 +42,7 @@ test-case = "3.1.0"
 
 [build-dependencies]
 eyre = "0.6.8"
+md5 = "0.7.0"
 rsonpath-test-codegen = { path = "../rsonpath-test-codegen" }
 
 [features]
