+++ crates/rsonpath-lib/tests/documents/json/large/twitter.json
