+++ crates/rsonpath-lib/tests/documents/toml/empty_object.toml
