+++ crates/rsonpath-lib/tests/documents/toml/singletons_and_empties.toml
