+++ crates/rsonpath-lib/tests/documents/toml/spaced_colon.toml
