+++ crates/rsonpath-test-codegen/Cargo.toml
@@ -12,8 +12,7 @@ categories = [
     "text-processing",
 ]
 repository = "https://github.com/V0ldek/rsonpath"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+rust-version = "1.70.0"
 
 [dependencies]
 serde = { version = "1.0.164", features = ["derive"] }
