--- crates/rsonpath-test/Cargo.toml
@@ -1,31 +1,0 @@
-[package]
-name = "rsonpath-test"
-version = "0.5.0"
-authors = ["Mateusz Gienieczko <mat@gienieczko.com>"]
-edition = "2021"
-description = "Blazing fast JSONPath query engine powered by SIMD. E2E test suite for rsonpath-lib."
-readme = "README.md"
-license = "MIT"
-keywords = ["json", "jsonpath", "query", "simd", "parser"]
-exclude = ["/tests"]
-categories = [
-    "text-processing",
-]
-repository = "https://github.com/V0ldek/rsonpath"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dev-dependencies]
-pretty_assertions = "1.3.0"
-rsonpath-lib = { version = "0.5.0", default-features = false }
-tempfile = "3.6.0"
-
-[build-dependencies]
-eyre = "0.6.8"
-md5 = "0.7.0"
-rsonpath-test-codegen = { path = "../rsonpath-test-codegen" }
-
-[features]
-default = ["simd", "unique-members"]
-simd = ["rsonpath-lib/simd"]
-unique-members = ["rsonpath-lib/unique-members"]
