--- crates/rsonpath-test/README.md
@@ -1,1 +1,0 @@
-wait for it
